+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+dev:
+  - add "proposer duties"
+
 1.24.1:
   - fix potential crash when new validators are activated
   - add "sepolia" to the list of supported networks